Ensure proper initialization of derived model states in mj_data when adding new models.
Call mj_forward to calculate forward kinematics / dynamics without progressing the simulation time. This ensures that all xpos, xquat, xmat properties of bodies, sites, etc. in the mj_data array are calculated from the actual simulation state after adding new robots.

An explicit initialization of the actual model state for actuators of a newly added robot model is still missing. Mujoco allows the specification of key-frames for a model. Maybe we can use this feature to initialize a robot in a specific posture.