Added vision sensor-pipeline implementation.
* Added perceptor for vision detections.
* Added point detection and player detection (represented by multiple point detections).
* Added generation of vision-pipeline perceptions to simulation server.
* Added visual marker and camera site specifications to soccer world and and robot model.

With this commit the server provides a vision-pipeline perception, which contains a list of vision detections of general world objects and agent body parts. In general, the vision-pipeline detects arbitrary site elements of world and robot specifications that are marked with a "-vismarker" suffix. This way, each world and robot model can specify its own set of visual markers that are detected by the vision-pipeline.
So far, only point detections are implemented. Calculations for line-segment detections are more involved due to the fact that they can be partially visible. The protocol again replicates the one from SimSpark for client software compatibility reasons.