Separated simulation and server components.
* Separated simulation and server components into different classes. The simulation server now uses a simulation instance to run the simulation, rather than extending a base simulation class.
* Removed managed server.
* Moved as much functionality as possible into the base simulation class. Introduced some abstract methods to deal with domain specific logic.
* Added soccer specific simulation.
* Inject parsers and encoders into remote client and monitor instances.

This is the first commit towards a generic simulation environment.
Simulation specific message parsers and encoders are already implemented.
There still is some domain specific dependency to the soccer referee which has to be eliminated.
Apart from that, the internal monitor is currently also domain specific due to displaying the soccer game state.